cleanup scikit class
