+++ marginaleffects/predictions.py
@@ -1,7 +1,6 @@
 import numpy as np
 import patsy
 import polars as pl
-import formulaic
 
 from .by import get_by
 from .classes import MarginaleffectsDataFrame
@@ -16,7 +15,7 @@ from .sanity import (
 )
 from .transform import get_transform
 from .uncertainty import get_jacobian, get_se, get_z_p_ci
-from .utils import sort_columns, ingest
+from .utils import sort_columns
 from .model_pyfixest import ModelPyfixest
 
 
@@ -134,7 +133,11 @@ def predictions(
         exog = newdata.to_pandas()
     else:
         if hasattr(model, "formula") and hasattr(model, "data"):
-            endog, exog = formulaic.model_matrix(model.formula, model.data.to_pandas())
+            try:
+                import formulaic
+                endog, exog = formulaic.model_matrix(model.formula, model.data.to_pandas())
+            except ImportError:
+                raise ImportError("The formulaic package is required to use this feature.")
         else:
             design_info = model.model.model.data.design_info
             exog = patsy.dmatrix(design_info, newdata.to_pandas(), NA_action="raise")
