+++ marginaleffects/utils.py
@@ -1,3 +1,4 @@
+import re
 import itertools
 import narwhals as nw
 import numpy as np
@@ -128,3 +129,26 @@ def get_type_dictionary(modeldata):
         else:
             out[v] = "unknown"
     return out
+
+
+def formula_to_variables(formula, newdata):
+    columns = newdata.columns
+    order = {}
+    for var in columns:
+        match = re.search(rf"\b{re.escape(var)}\b", formula.split("~")[1])
+        if match:
+            order[var] = match.start()
+    variables = sorted(order, key=lambda i: order[i])
+
+    if isinstance(variables, (str, dict)):
+        variables = [variables] if isinstance(variables, str) else variables
+    elif isinstance(variables, list) and all(
+        isinstance(var, str) for var in variables
+    ):
+        pass
+    else:
+        raise ValueError(
+            "`variables` must be None, a dict, string, or list of strings"
+        )
+
+    return variables
