+++ pyproject.toml
@@ -16,6 +16,7 @@ dependencies = [
 [project.optional-dependencies]
 test = [
     "duckdb >=1.1.2",
+    "formulaic >=1.0.2",
     "matplotlib >=3.7.1",
     "pandas >=2.2.2",
     "pyarrow >=17.0.0",
